feat(report-cards): migrate basal vs_literature to Gen 2 VsLiteratureCard
Move the basal `vs_literature` report card off the standalone render script
onto the Gen 2 per-study `ReportCardStep` surface, so it's dashboard-native
and flushed like every other card (tests, vs_vecoli).

- v2ecoli/library/vs_literature.py: lift the reusable, import-safe core
  (reference-building + grading inputs; reads the committed
  tests/fixtures/population_phenotype_basal/model_*.json and the ecoli-sources
  validation bundle). The heavy sweep-bake path + CLI stay in the script.
- scripts/render_basal_vs_literature.py: thin CLI wrapper that re-exports the
  library names, so tests/test_basal_vs_literature.py keeps importing them.
- v2ecoli/workflow/report_cards/vs_literature_card.py: VsLiteratureCard
  (name="vs_literature"), auto-registers in REPORT_CARD_REGISTRY +
  POST_SIM_REGISTRY. applies() gates on report_card_refs.vs_literature +
  resolvable fixtures (returns False cleanly on other studies). build()
  grades via grade_card and renders the rich HTML (render_html), matching the
  standalone card's output.
- workspace/studies/population_phenotype_basal/study.yaml: schema-v4 study so
  the card attaches on the dashboard; scoped to the vs_literature card.
- tests/test_vs_literature_card.py: registration/applies/build + the migration
  gate (verdict identical to the pre-migration test_basal_vs_literature golden).

Behavior-preserving: the runner writes viz/report_card/vs_literature.{html,
verdict.json}, and the verdict is identical to the pre-migration golden
(13/13 axes, overall mismatch). Measured side is the committed fixtures (the
#268 reproducibility boundary), not a live zarr — reversible to a later
study.run_zarr_paths() drop-in.
