+++ v2ecoli/workflow/report_cards/__init__.py
@@ -156,7 +156,8 @@ def applicable(ctx: StudyContext, core, only: "str | None" = None) -> list:
     return out
 
 
-# Register built-in cards (import for side effect). Both modules exist, so
+# Register built-in cards (import for side effect). These modules all exist, so
 # import unconditionally — a real import error must surface, not be masked.
 from . import tests_card  # noqa: E402,F401
+from . import vs_literature_card  # noqa: E402,F401
 from . import vs_vecoli_card  # noqa: E402,F401
